ml_gemini_chatbot/requirements.txt
pandas
scikit-learn
joblib
streamlit
google-generativeai
flask
python-dotenv
matplotlib  x   